### 2025-10-18 #9273 @editorialbot
[PRE REVIEW]: pyMultiFit: A Python library for fitting data with multiple models
**Submitting author:** @syedalimohsinbukhari (<a href="http://orcid.org/0000-0002-4232-3874">Syed Ali Mohsin Bukhari</a>)
**Repository:** https://github.com/syedalimohsinbukhari/pyMultiFit
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.7
