### 2025-10-19 #9273 @syedalimohsinbukhari
Hi @crvernon thanks for pointing out the issue, I've fixed it.

I'd also like to point out that I had submitted `pyMultiFit` for review in `PyOpenSci` [Issue 233](https://github.com/pyOpenSci/software-submission/issues/233)
