### 2025-10-23 #9295 @arfon
Moving forward without additional review due to completed pyOpenSci review in https://github.com/pyOpenSci/software-submission/issues/233
