### 2025-10-24 #9295 @arfon
Apologies @syedalimohsinbukhari – looks like the figure labels are now duplicative. This PR should fix it: https://github.com/syedalimohsinbukhari/pyMultiFit/pull/106
