### 2025-10-24 #9295 @syedalimohsinbukhari
@arfon merged,
