### 2025-12-24 #9295 @danielskatz
The source for your paper is https://github.com/syedalimohsinbukhari/pyMultiFit/blob/main/inst/paper.md isn't it?

What do you mean by the review branch? The JOSS review looked at the main branch, and I don't see any branch named review, other than "pyopensci-review", which isn't JOSS
